+++ src/utils/common/netio.c
@@ -321,7 +321,7 @@ int net_send(const net_t *net, const uint8_t *buf, const size_t buf_len)
 	} else {
 		// Send data.
 		if (sendto(net->sockfd, buf, buf_len, 0, net->srv->ai_addr,
-		           net->srv->ai_addrlen) != buf_len) {
+		           net->srv->ai_addrlen) != (ssize_t)buf_len) {
 			WARN("can't send query to %s\n", net->remote_str);
 			return KNOT_NET_ESEND;
 		}
